+++ barely/core/ChangeTracker.py
@@ -41,7 +41,7 @@ class ChangeTracker:
         ignore_patterns = ""
         ignore_directories = False
         case_sensitive = True
-        handler = PatternMatchingEventHandler(patterns, ignore_patterns, ignore_directories, case_sensitive)
+        handler = PatternMatchingEventHandler(patterns=patterns, ignore_patterns=ignore_patterns, ignore_directories=ignore_directories, case_sensitive=case_sensitive)
 
         handler.on_created = self.buffer
         handler.on_deleted = self.buffer